Removed abstract model
